Enhanced category management and spotlight integration:
- Added support for `image` field in `CategoryReq` and `Category` entity for category images.
- Updated `AdminCategoryRes` to include `imageKey`, `spotlightName`, and reordered fields for clarity.
- Refined `updateCategoryOrder` method and controller to use `Long` for order values.
- Improved HTTP response consistency across admin and spotlight controllers.
- Streamlined category-related methods in `CategoryServiceImpl` with additional spotlight and image handling.
- Integrated `AttachmentRepository` for managing category images.

+++ src/main/java/org/exp/primeapp/service/interfaces/user/CategoryService.java
@@ -29,7 +29,7 @@ public interface CategoryService {
 
     void toggleCategoryActiveStatusWithProductActiveStatus(Long categoryId);
 
-    List<AdminCategoryRes> updateCategoryOrder(Map<Long, Integer> categoryOrderMap);
+    List<AdminCategoryRes> updateCategoryOrder(Map<Long, Long> categoryOrderMap);
 
     List<CategoryRes> getSpotlightCategories(Long spotlightId);
 }
